PR libfortran/32021 	* runtime/backtrace.c (local_strcasestr): Protect by appropriate 	macros. 	* runtime/main.c (cleanup): Cast argument to free. 	* intrinsics/spread_generic.c (spread_internal): Match runtime_error 	arguments and format. 	* intrinsics/signal.c (alarm_sub_int_i4, alarm_sub_int_i8): Cast 	pointers to avoid warnings.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@129463 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgfortran/intrinsics/spread_generic.c
@@ -131,9 +131,9 @@ spread_internal (gfc_array_char *ret, const gfc_array_char *source,
 
 		  if (ret_extent != ncopies)
 		    runtime_error("Incorrect extent in return value of SPREAD"
-				  " intrinsic in dimension %d: is %ld,"
-				  " should be %ld", n+1, (long int) ret_extent,
-				  (long int) ncopies);
+				  " intrinsic in dimension %ld: is %ld,"
+				  " should be %ld", (long int) n+1,
+				  (long int) ret_extent, (long int) ncopies);
 		}
 	      else
 		{
@@ -142,8 +142,9 @@ spread_internal (gfc_array_char *ret, const gfc_array_char *source,
 		    - source->dim[dim].lbound;
 		  if (ret_extent != extent[dim])
 		    runtime_error("Incorrect extent in return value of SPREAD"
-				  " intrinsic in dimension %d: is %ld,"
-				  " should be %ld", n+1, (long int) ret_extent,
+				  " intrinsic in dimension %ld: is %ld,"
+				  " should be %ld", (long int) n+1,
+				  (long int) ret_extent,
 				  (long int) extent[dim]);
 		    
 		  if (extent[dim] <= 0)
